+++ engine/src/main/java/org/camunda/bpm/engine/impl/cfg/ProcessEngineConfigurationImpl.java
@@ -1021,6 +1021,7 @@ public abstract class ProcessEngineConfigurationImpl extends ProcessEngineConfig
   // diagnostics ///////////////////////////////////////////////////////
   protected DiagnosticsCollector diagnosticsCollector;
   protected TelemetryDataImpl telemetryData;
+  protected DiagnosticsRegistry diagnosticsRegistry;
 
   // Exception Codes ///////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -5292,4 +5293,12 @@ public abstract class ProcessEngineConfigurationImpl extends ProcessEngineConfig
     return new ExceptionCodeInterceptor(builtinExceptionCodeProvider, customExceptionCodeProvider);
   }
 
+  public DiagnosticsRegistry getDiagnosticsRegistry() {
+    return diagnosticsRegistry;
+  }
+
+  public ProcessEngineConfiguration setDiagnosticsRegistry(DiagnosticsRegistry diagnosticsRegistry) {
+    this.diagnosticsRegistry = diagnosticsRegistry;
+    return this;
+  }
 }
