+++ spring-boot-starter/starter-qa/integration-test-simple/src/test/java/org/camunda/bpm/springboot/project/qa/simple/SimpleApplicationIT.java
@@ -20,6 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.camunda.bpm.engine.ProcessEngine;
 import org.camunda.bpm.engine.RuntimeService;
+import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.camunda.bpm.engine.impl.diagnostics.DiagnosticsRegistry;
 import org.camunda.bpm.engine.impl.telemetry.dto.ApplicationServerImpl;
 import org.junit.Test;
@@ -52,7 +53,7 @@ public class SimpleApplicationIT {
   @Test
   public void shouldNotDetermineApplicationServer() {
 
-    DiagnosticsRegistry diagnosticsRegistry = processEngine.getProcessEngineConfiguration().getDiagnosticsRegistry();
+    DiagnosticsRegistry diagnosticsRegistry = ((ProcessEngineConfigurationImpl) processEngine.getProcessEngineConfiguration()).getDiagnosticsRegistry();
 
     // then
     ApplicationServerImpl applicationServer = diagnosticsRegistry.getApplicationServer();
