+++ spring-boot-starter/starter/src/test/java/org/camunda/bpm/spring/boot/starter/TelemetryNonPaIT.java
@@ -41,7 +41,7 @@ public class TelemetryNonPaIT extends AbstractCamundaAutoConfigurationIT {
 
   @Test
   public void shouldSubmitApplicationServerData() {
-    DiagnosticsRegistry diagnosticsRegistry = processEngine.getProcessEngineConfiguration().getDiagnosticsRegistry();
+    DiagnosticsRegistry diagnosticsRegistry = ((ProcessEngineConfigurationImpl) processEngine.getProcessEngineConfiguration()).getDiagnosticsRegistry();
 
     // then
     ApplicationServerImpl applicationServer = diagnosticsRegistry.getApplicationServer();
